refactor code and make bet logic

+++ backend/src/main/java/ru/netcracker/backend/requests/CurrencyRequest.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class CurrencyRequest {
     private String username;
-    private Long currency;
+    private BigDecimal currency;
 }
