+++ backend/src/main/java/ru/netcracker/backend/responses/UserResponse.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -16,5 +18,5 @@ public class UserResponse {
     private String about;
     private Boolean isBanned;
     private Boolean enabled;
-    private long currency;
+    private BigDecimal currency;
 }
